refactor remove_orphaned
